Move all fixtures to conftest.py
